+++ src/transport/tcp/tcp.c
@@ -115,7 +115,7 @@ nni_tcp_pipe_recv(void *arg, nni_msg **msgp)
 		return (NNG_EPROTO);
 	}
 
-	if ((rv = nng_msg_alloc(&msg, len)) != 0) {
+	if ((rv = nng_msg_alloc(&msg, (size_t) len)) != 0) {
 		return (rv);
 	}
